feat(beatstep): sleep 1ms between sysex commands
When SysEx commands are sent too quickly the beatstep will skip some
commands
and end up with completely and randomly scrambled lights.

This commit introduces a 1ms `Thread.sleep` after each SysEx command
which is not particularly elegant but fixes the issue.
I have, so far, not encountered any missing note inputs or such because
of the sleeping thread so this is probably fine.